Previous removal of fnc were not updated in all files
Changes:
The previous code removed fnc from the method names but did not update it through all the files so the code wouldn't run. This is now fixed.
The processed data is now stored in a tempdata folder that can now make it easier to delete the data that's not needed anymore. The images of the graphs are now saved in the date time stamp folder.

Bugs: 
Can't download full globe for 8 days, file size way too big and server doesn't allow either
The parameters seem to be off, when plotting the graphs look really odd. Not sure what changed, if it was a problem with the data or something else.